In root package routes also renamed to template. Jinja2 templates stored in settings
